move phone-api to own dir
