+++ api/phone-api/server/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "server",
     srcs = ["server.go"],
-    importpath = "github.com/rickypai/web-template/api/server",
+    importpath = "github.com/rickypai/web-template/api/phone-api/server",
     visibility = ["//visibility:public"],
     deps = [
         "//api/server/cursor",
